fix: query classes made static and naming && null-check fixed
