+++ Source/Iwentys.EntityManager.WebApi/Controllers/StudyGroups/Queries/GetStudyGroupByGroupName.cs
@@ -1,16 +1,17 @@
 ﻿using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Iwentys.EntityManager.DataAccess;
+using Iwentys.EntityManager.Domain;
 using Iwentys.EntityManager.WebApiDtos;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
 namespace Iwentys.EntityManager.WebApi;
 
-public class GetStudyGroupByCourseId
+public static class GetStudyGroupByGroupName
 {
-    public record Query(int? CourseId) : IRequest<Response>;
-    public record Response(List<StudyGroupProfileResponseDto> StudyGroups);
+    public record Query(string GroupName) : IRequest<Response>;
+    public record Response(StudyGroupProfileResponseDto StudyGroup);
 
     public class Handler : IRequestHandler<Query, Response>
     {
@@ -25,11 +26,12 @@ public class GetStudyGroupByCourseId
 
         public async Task<Response> Handle(Query request, CancellationToken cancellationToken)
         {
-            List<StudyGroupProfileResponseDto> result = await _context
+            var name = new GroupName(request.GroupName);
+            var result = await _context
                 .StudyGroups
-                .WhereIf(request.CourseId, gs => gs.StudyCourseId == request.CourseId)
+                .Where(StudyGroup.IsMatch(name))
                 .ProjectTo<StudyGroupProfileResponseDto>(_mapper.ConfigurationProvider)
-                .ToListAsync(cancellationToken: cancellationToken);
+                .SingleOrDefaultAsync(cancellationToken: cancellationToken);
 
             return new Response(result);
         }
