+++ Source/Iwentys.EntityManager.WebApi/Controllers/StudyGroups/StudyGroupController.cs
@@ -22,50 +22,46 @@ public class StudyGroupController : ControllerBase
         return Ok(response.StudyGroups);
     }
 
+    [HttpGet(nameof(GetStudyGroupByGroupName))]
+    public async Task<ActionResult<StudyGroupProfileResponseDto>> GetStudyGroupByGroupName(string groupName)
+    {
+        GetStudyGroupByGroupName.Response response = await _mediator.Send(new GetStudyGroupByGroupName.Query(groupName));
+        var result = response?.StudyGroup;
+
+        return result is not null ? Ok(result) : NotFound();
+    }
+
+    [HttpGet(nameof(GetStudyGroupByStudentId))]
+    public async Task<ActionResult<StudyGroupProfileResponseDto>> GetStudyGroupByStudentId(int studentId)
+    {
+        GetStudyGroupByStudentId.Response response = await _mediator.Send(new GetStudyGroupByStudentId.Query(studentId));
+        var result = response?.StudyGroup;
+
+        return result is not null ? Ok(result) : NotFound();
+    }
+
     [HttpGet(nameof(GetStudyGroupById))]
     public async Task<ActionResult<StudyGroupProfileResponseDto>> GetStudyGroupById(int groupId)
     {
         GetStudyGroupById.Response response = await _mediator.Send(new GetStudyGroupById.Query(groupId));
+        var result = response?.StudyGroup;
 
-        if (response?.StudyGroup is null) return NotFound();
-
-        return Ok(response.StudyGroup);
+        return result is not null ? Ok(result) : NotFound();
     }
 
     [HttpPost(nameof(GetStudyGroupsByIdList))]
     public async Task<ActionResult<IReadOnlyCollection<StudyGroupProfileResponseDto>>> GetStudyGroupsByIdList(
         List<int> groupIdList)
     {
-        GetStudyGroupsByIdList.Response response
-            = await _mediator.Send(new GetStudyGroupsByIdList.Query(groupIdList));
+        GetStudyGroupsByIdList.Response response = await _mediator.Send(new GetStudyGroupsByIdList.Query(groupIdList));
 
         return Ok(response.StudyGroups);
     }
 
     [HttpGet(nameof(GetStudyGroupsByCourseId))]
-    public async Task<ActionResult<IReadOnlyCollection<StudyGroupProfileResponseDto>>> GetStudyGroupsByCourseId(int? courseId)
+    public async Task<ActionResult<IReadOnlyCollection<StudyGroupProfileResponseDto>>> GetStudyGroupsByCourseId(int courseId)
     {
-        GetStudyGroupByCourseId.Response response = await _mediator.Send(new GetStudyGroupByCourseId.Query(courseId));
+        GetStudyGroupsByCourseId.Response response = await _mediator.Send(new GetStudyGroupsByCourseId.Query(courseId));
         return Ok(response.StudyGroups);
-    }
-
-    [HttpGet(nameof(GetStudyGroupByGroupName))]
-    public async Task<ActionResult<StudyGroupProfileResponseDto>> GetStudyGroupByGroupName(string groupName)
-    {
-        GetStudyGroupByName.Response response = await _mediator.Send(new GetStudyGroupByName.Query(groupName));
-
-        if (response?.StudyGroup is null) return NotFound();
-
-        return Ok(response.StudyGroup);
-    }
-
-    [HttpGet(nameof(GetStudyGroupByStudentId))]
-    public async Task<ActionResult<StudyGroupProfileResponseDto>> GetStudyGroupByStudentId(int studentId)
-    {
-        GetStudyGroupByStudent.Response response = await _mediator.Send(new GetStudyGroupByStudent.Query(studentId));
-
-        if (response?.StudyGroup is null) return NotFound();
-
-        return Ok(response.StudyGroup);
     }
 }
